+++ toolkit/include/toolkit/box-algorithms/ovtkTBoxAlgorithm.h
@@ -340,24 +340,24 @@ namespace OpenViBEToolkit
 		virtual bool onNameChanged(OpenViBE::Kernel::IBox& /*box*/) { return true; }
 		virtual bool onIdentifierChanged(OpenViBE::Kernel::IBox& /*box*/) { return true; }
 		virtual bool onAlgorithmClassIdentifierChanged(OpenViBE::Kernel::IBox& /*box*/) { return true; }
-		virtual bool onInputConnected(OpenViBE::Kernel::IBox& /*box*/, const uint32_t /*index*/) { return true; }
-		virtual bool onInputDisconnected(OpenViBE::Kernel::IBox& /*box*/, const uint32_t /*index*/) { return true; }
-		virtual bool onInputAdded(OpenViBE::Kernel::IBox& /*box*/, const uint32_t /*index*/) { return true; }
-		virtual bool onInputRemoved(OpenViBE::Kernel::IBox& /*box*/, const uint32_t /*index*/) { return true; }
-		virtual bool onInputTypeChanged(OpenViBE::Kernel::IBox& /*box*/, const uint32_t /*index*/) { return true; }
-		virtual bool onInputNameChanged(OpenViBE::Kernel::IBox& /*box*/, const uint32_t /*index*/) { return true; }
-		virtual bool onOutputConnected(OpenViBE::Kernel::IBox& /*box*/, const uint32_t /*index*/) { return true; }
-		virtual bool onOutputDisconnected(OpenViBE::Kernel::IBox& /*box*/, const uint32_t /*index*/) { return true; }
-		virtual bool onOutputAdded(OpenViBE::Kernel::IBox& /*box*/, const uint32_t /*index*/) { return true; }
-		virtual bool onOutputRemoved(OpenViBE::Kernel::IBox& /*box*/, const uint32_t /*index*/) { return true; }
-		virtual bool onOutputTypeChanged(OpenViBE::Kernel::IBox& /*box*/, const uint32_t /*index*/) { return true; }
-		virtual bool onOutputNameChanged(OpenViBE::Kernel::IBox& /*box*/, const uint32_t /*index*/) { return true; }
-		virtual bool onSettingAdded(OpenViBE::Kernel::IBox& /*box*/, const uint32_t /*index*/) { return true; }
-		virtual bool onSettingRemoved(OpenViBE::Kernel::IBox& /*box*/, const uint32_t /*index*/) { return true; }
-		virtual bool onSettingTypeChanged(OpenViBE::Kernel::IBox& /*box*/, const uint32_t /*index*/) { return true; }
-		virtual bool onSettingNameChanged(OpenViBE::Kernel::IBox& /*box*/, const uint32_t /*index*/) { return true; }
-		virtual bool onSettingDefaultValueChanged(OpenViBE::Kernel::IBox& /*box*/, const uint32_t /*index*/) { return true; }
-		virtual bool onSettingValueChanged(OpenViBE::Kernel::IBox& /*box*/, const uint32_t /*index*/) { return true; }
+		virtual bool onInputConnected(OpenViBE::Kernel::IBox& /*box*/, const size_t /*index*/) { return true; }
+		virtual bool onInputDisconnected(OpenViBE::Kernel::IBox& /*box*/, const size_t /*index*/) { return true; }
+		virtual bool onInputAdded(OpenViBE::Kernel::IBox& /*box*/, const size_t /*index*/) { return true; }
+		virtual bool onInputRemoved(OpenViBE::Kernel::IBox& /*box*/, const size_t /*index*/) { return true; }
+		virtual bool onInputTypeChanged(OpenViBE::Kernel::IBox& /*box*/, const size_t /*index*/) { return true; }
+		virtual bool onInputNameChanged(OpenViBE::Kernel::IBox& /*box*/, const size_t /*index*/) { return true; }
+		virtual bool onOutputConnected(OpenViBE::Kernel::IBox& /*box*/, const size_t /*index*/) { return true; }
+		virtual bool onOutputDisconnected(OpenViBE::Kernel::IBox& /*box*/, const size_t /*index*/) { return true; }
+		virtual bool onOutputAdded(OpenViBE::Kernel::IBox& /*box*/, const size_t /*index*/) { return true; }
+		virtual bool onOutputRemoved(OpenViBE::Kernel::IBox& /*box*/, const size_t /*index*/) { return true; }
+		virtual bool onOutputTypeChanged(OpenViBE::Kernel::IBox& /*box*/, const size_t /*index*/) { return true; }
+		virtual bool onOutputNameChanged(OpenViBE::Kernel::IBox& /*box*/, const size_t /*index*/) { return true; }
+		virtual bool onSettingAdded(OpenViBE::Kernel::IBox& /*box*/, const size_t /*index*/) { return true; }
+		virtual bool onSettingRemoved(OpenViBE::Kernel::IBox& /*box*/, const size_t /*index*/) { return true; }
+		virtual bool onSettingTypeChanged(OpenViBE::Kernel::IBox& /*box*/, const size_t /*index*/) { return true; }
+		virtual bool onSettingNameChanged(OpenViBE::Kernel::IBox& /*box*/, const size_t /*index*/) { return true; }
+		virtual bool onSettingDefaultValueChanged(OpenViBE::Kernel::IBox& /*box*/, const size_t /*index*/) { return true; }
+		virtual bool onSettingValueChanged(OpenViBE::Kernel::IBox& /*box*/, const size_t /*index*/) { return true; }
 
 		// ====================================================================================================================================
